+++ Source/ExecutionEngine/ExecutionEngine.cs
@@ -1310,6 +1310,7 @@ namespace Microsoft.Boogie
         return;
       }
 
+      errors = errors.DistinctBy(e => e.AssertCmd).ToList();
       errors.Sort(new CounterexampleComparer());
       foreach (Counterexample error in errors)
       {
